添加假数据

+++ src/main/java/com/bjlx/QinShihuang/model/poi/Restaurant.java
@@ -148,13 +148,13 @@ public class Restaurant {
 	 * POI价格
 	 */
 	@NotBlank
-	private Double price = 0.0;
+	private Integer price = 0;
 	
 	/**
 	 * POI价格
 	 */
 	@NotBlank
-	private Double marketPrice = 0.0;
+	private Integer marketPrice = 0;
 	
 	/**
 	 * POI价格描述
@@ -322,11 +322,11 @@ public class Restaurant {
 		this.url = url;
 	}
 
-	public Double getPrice() {
+	public Integer getPrice() {
 		return price;
 	}
 
-	public void setPrice(Double price) {
+	public void setPrice(Integer price) {
 		this.price = price;
 	}
 
@@ -426,11 +426,11 @@ public class Restaurant {
 		this.saleVolume = saleVolume;
 	}
 
-	public Double getMarketPrice() {
+	public Integer getMarketPrice() {
 		return marketPrice;
 	}
 
-	public void setMarketPrice(Double marketPrice) {
+	public void setMarketPrice(Integer marketPrice) {
 		this.marketPrice = marketPrice;
 	}
 
@@ -446,7 +446,7 @@ public class Restaurant {
 
 	}
 
-	public Restaurant(String id, ImageItem cover, String zhName, String enName, String url, Double price, Double marketPrice) {
+	public Restaurant(String id, ImageItem cover, String zhName, String enName, String url, Integer price, Integer marketPrice) {
 		this.id = new ObjectId(id);
 		this.cover = cover;
 		this.zhName = zhName;
