+++ src/main/java/com/bjlx/QinShihuang/model/poi/Shopping.java
@@ -152,13 +152,13 @@ public class Shopping {
 	 * POI价格
 	 */
 	@NotBlank
-	private Double price = 0.0;
+	private Integer price = 0;
 	
 	/**
 	 * POI市场价格
 	 */
 	@NotBlank
-	private Double marketPrice = 0.0;
+	private Integer marketPrice = 0;
 	
 	/**
 	 * POI价格描述
@@ -331,11 +331,11 @@ public class Shopping {
 		this.url = url;
 	}
 
-	public Double getPrice() {
+	public Integer getPrice() {
 		return price;
 	}
 
-	public void setPrice(Double price) {
+	public void setPrice(Integer price) {
 		this.price = price;
 	}
 
@@ -427,11 +427,11 @@ public class Shopping {
 		this.locality = locality;
 	}
 
-	public Double getMarketPrice() {
+	public Integer getMarketPrice() {
 		return marketPrice;
 	}
 
-	public void setMarketPrice(Double marketPrice) {
+	public void setMarketPrice(Integer marketPrice) {
 		this.marketPrice = marketPrice;
 	}
 
@@ -463,7 +463,7 @@ public class Shopping {
 
 	}
 
-	public Shopping(String id, ImageItem cover, String zhName, String enName, String url, Double price, Double marketPrice) {
+	public Shopping(String id, ImageItem cover, String zhName, String enName, String url, Integer price, Integer marketPrice) {
 		this.id = new ObjectId(id);
 		this.cover = cover;
 		this.zhName = zhName;
