+++ src/main/java/com/bjlx/QinShihuang/requestmodel/RestaurantReq.java
@@ -45,13 +45,13 @@ public class RestaurantReq {
      * POI价格
      */
     @Min(value = 0)
-    private Double price = 0.0;
+    private Integer price = 0;
 
     /**
      * POI价格
      */
     @Min(value = 0)
-    private Double marketPrice = 0.0;
+    private Integer marketPrice = 0;
 
 
     public String getId() {
@@ -94,19 +94,19 @@ public class RestaurantReq {
         this.url = url;
     }
 
-    public Double getPrice() {
+    public Integer getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(Integer price) {
         this.price = price;
     }
 
-    public Double getMarketPrice() {
+    public Integer getMarketPrice() {
         return marketPrice;
     }
 
-    public void setMarketPrice(Double marketPrice) {
+    public void setMarketPrice(Integer marketPrice) {
         this.marketPrice = marketPrice;
     }
 }
