+++ src/main/java/com/bjlx/QinShihuang/requestmodel/ShoppingReq.java
@@ -45,13 +45,13 @@ public class ShoppingReq {
      * POI价格
      */
     @Min(value = 0)
-    private Double price = 0.0;
+    private Integer price = 0;
 
     /**
      * POI市场价格
      */
     @Min(value = 0)
-    private Double marketPrice = 0.0;
+    private Integer marketPrice = 0;
 
     public String getId() {
         return id;
@@ -93,19 +93,19 @@ public class ShoppingReq {
         this.url = url;
     }
 
-    public Double getPrice() {
+    public Integer getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(Integer price) {
         this.price = price;
     }
 
-    public Double getMarketPrice() {
+    public Integer getMarketPrice() {
         return marketPrice;
     }
 
-    public void setMarketPrice(Double marketPrice) {
+    public void setMarketPrice(Integer marketPrice) {
         this.marketPrice = marketPrice;
     }
 }
