+++ src/main/java/com/bjlx/QinShihuang/requestmodel/ViewspotReq.java
@@ -44,13 +44,13 @@ public class ViewspotReq {
      * POI价格
      */
     @Min(value = 0)
-    private Double price = 0.0;
+    private Integer price = 0;
 
     /**
      * 市场价
      */
     @Min(value = 0)
-    private Double marketPrice = 0.0;
+    private Integer marketPrice = 0;
 
     public String getId() {
         return id;
@@ -92,19 +92,19 @@ public class ViewspotReq {
         this.url = url;
     }
 
-    public Double getPrice() {
+    public Integer getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(Integer price) {
         this.price = price;
     }
 
-    public Double getMarketPrice() {
+    public Integer getMarketPrice() {
         return marketPrice;
     }
 
-    public void setMarketPrice(Double marketPrice) {
+    public void setMarketPrice(Integer marketPrice) {
         this.marketPrice = marketPrice;
     }
 }
